+++ NEWS.md
@@ -1,3 +1,10 @@
+# roz 0.0.3
+
+* Add `tools/exercise_roz_api.R` for testing the interface between `roz` and the OZ API.
+* Remove return value from `oz_basereq()`.
+* Force ott ids to be returned by `node_images()` as integers.
+* Mitigate issue with header values from the OneZoom popularity endpoint. See [Issue #875](https://github.com/OneZoom/OZtree/issues/875) for details.
+
 # roz 0.0.2
 
 * Add `node_images()` to allow for retrieval of public domain and cc images.